ai/ImageAI/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
azure-storage-blob
python-multipart
opencv-python-headless
numpy
httpx==0.24.1